chore: rename RedirectManager to ComponentWithRedirectManager
