+++ packages/sdk/src/HeadlessUniversalCheckout/Managers/ComponentWithRedirectManager.ts
